pyproject.toml
[project]
name = "nexustrader-mcp"
version = "0.1.0"
description = "MCP server for NexusTrader — expose trading, market data, and account tools to AI clients"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "nexustrader",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
]

[project.scripts]
nexustrader-mcp = "nexustrader_mcp.cli:cli_entry"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
nexustrader = { path = "../NexusTrader", editable = true }
